fix(trip): refactor trip domain imports
