+++ src/package/trip/application/get-trip-by-status.ts
@@ -3,7 +3,7 @@ import {
   Ok,
   Result
 } from 'oxide.ts'
-import { TripState } from 'src/app/shared/models/trip-state'
+import { TripState } from 'src/package/trip/domain/models/trip-state'
 import { TripDao } from 'src/package/trip/domain/dao/trip-dao'
 import { Trip } from 'src/package/trip/domain/models/trip'
