+++ src/package/trip/domain/dao/trip-dao.ts
@@ -1,5 +1,5 @@
 import { Result } from 'oxide.ts'
-import { TripState } from 'src/app/shared/models/trip-state'
+import { TripState } from 'src/package/trip/domain/models/trip-state'
 import { Trip } from 'src/package/trip/domain/models/trip'
 import { TripID } from 'src/package/trip/domain/models/trip-id'
