+++ src/package/trip/domain/models/trip-price.ts
