+++ src/package/trip/domain/models/trip-seat.ts
