+++ src/package/trip/domain/models/trip-state.ts
